[WARP-SPEED] copy RecordIndexes to loaders
(cherry picked from commit 3135519bd595d52ac4525e99547a2df674a37b7b)

+++ plugin/starburst-warp-speed/starburst-warp-speed-core/src/main/java/io/trino/plugin/warp/storage/read/MatcherArgs.java
@@ -20,7 +20,7 @@ import java.util.List;
 
 public record MatcherArgs(List<ReadJuffersWarmUpElement> matchJuffersWe,
                           LuceneMatcher[] luceneMatchers,
-                          byte[] storeRowListBuff,
+                          short[] storeRowListBuff,
                           ChunksQueue chunksQueue)
 {
 }
